feat: Conditional logging
