+++ googlepay/src/main/java/br/com/braspag/googlepay/BillingAddress.kt
@@ -1,0 +1,13 @@
+package br.com.braspag.googlepay
+
+data class BillingAddress(
+    val name: String,
+    val address1: String,
+    val address2: String,
+    val address3: String,
+    val locality: String,
+    val administrativeArea: String,
+    val postalCode: String,
+    val countryCode: String,
+    val sortingCode: String
+)
