+++ googlepay/src/main/java/br/com/braspag/googlepay/BraspagGooglePay.kt
@@ -3,6 +3,7 @@ package br.com.braspag.googlepay
 import android.app.Activity
 import android.content.Intent
 import android.util.Log
+import br.com.braspag.googlepay.common.*
 import br.com.braspag.googlepay.common.microsToString
 import br.com.braspag.googlepay.common.toBillingAddress
 import br.com.braspag.googlepay.common.toShippingAddress
@@ -33,10 +34,6 @@ class BraspagGooglePay(
     private val dataRequestCode: Int = 999
 ) {
 
-    companion object {
-        const val TAG = "BraspagGooglePay"
-    }
-
     private val googlePayHelper = GooglePayHelper()
 
     private lateinit var client: PaymentsClient
@@ -61,7 +58,8 @@ class BraspagGooglePay(
         val json = googlePayHelper.isReadyToPayRequest(billingAddressRequired)
         json?.let {
 
-            Log.d(TAG, "isReadyToPayRequest JSON: [$it]")
+
+            logd("isReadyToPayRequest JSON: [$it]")
             val request = IsReadyToPayRequest.fromJson(it.toString())
 
             val task = client.isReadyToPay(request)
@@ -69,7 +67,7 @@ class BraspagGooglePay(
             task.addOnCompleteListener { completedTask ->
                 try {
                     completedTask.getResult(ApiException::class.java)?.let { result ->
-                        Log.d(TAG, "result: [$result]")
+                        logd("result: [$result]")
                         callback.invoke(result)
                     } ?: callback.invoke(false)
                 } catch (exception: ApiException) {
@@ -96,7 +94,7 @@ class BraspagGooglePay(
         )
 
         if (json == null) {
-            Log.e("makeTransaction", "Can't fetch payment data request")
+            loge("Can't fetch payment data request")
             return
         }
 
@@ -121,7 +119,7 @@ class BraspagGooglePay(
         paymentData?.let {
             val paymentInformation = it.toJson()
 
-            Log.d(TAG, "paymentInformation : $paymentInformation")
+            logd( "paymentInformation : $paymentInformation")
 
             try {
                 val paymentMethodData =
@@ -146,15 +144,15 @@ class BraspagGooglePay(
 
                 val billingName = billingAddress?.getString("name")
 
-                Log.d(TAG, "BillingName : $billingName")
+                logd("BillingName : $billingName")
 
                 val token = paymentMethodData
                     .getJSONObject("tokenizationData")
                     .getString("token")
 
-                Log.d(TAG, "token: $token")
-                Log.d(TAG, "billingAddress: $billingAddress")
-                Log.d(TAG, "shippingAddress: $shippingAddress")
+                logd( "token: $token")
+                logd( "billingAddress: $billingAddress")
+                logd("shippingAddress: $shippingAddress")
 
                 val tokenObject = JSONObject(token)
 
@@ -169,7 +167,7 @@ class BraspagGooglePay(
                 )
 
             } catch (e: Throwable) {
-                Log.e(TAG, "Error: $e")
+                loge("Error: $e")
                 return null
             }
         }
