--- googlepay/src/main/java/br/com/braspag/googlepay/Constants.kt
@@ -1,6 +1,0 @@
-package br.com.braspag.googlepay
-
-object Constants {
-
-
-}
